folder rearragment, location resolvers
