+++ server/Category/CategoryResolvers.js
@@ -1,4 +1,4 @@
-const { Category } = require('../models/Category');
+const { Category } = require('./CategoryModel');
 
 const CategoryResolvers = {
   Query: {
