+++ server/Event/EventModel.js
@@ -14,6 +14,10 @@ const EventSchema = new Schema({
     type: String,
     require: true
   },
+  date: {
+    type: Date,
+    require: true
+  },
   attendees: [{
     type: mongoose.Schema.Types.ObjectId,
     ref: 'user'
